# server/src/http/routes/swagger.yml
openapi: 3.0.1
info:
  title: API Exposition InserJeunes
  description: |
    L’API Exposition InserJeunes permet **la récupération des [données InserJeunes](https://documentation.exposition.inserjeunes.beta.gouv.fr/)** (dont la “part en emploi à 6 mois” et le “taux en formation”).

    Les diverses routes permettent de récupérer ces données : 

    - pour une **formation / certification,** à partir d’un code MEFSTAT11, pour la voie scolaire, ou d’un code formation diplôme (CFD), pour la voie apprentissage.
    - pour une **formation / certification au sein d’un établissement,** à partir d’un code MEFSTAT11 ou CFD et d’un code UAI (Unité Administrative Immatriculée).

    L’API permet en outre de récupérer plusieurs de ces éléments à la fois, et propose divers filtres, par exemple pour récupérer les données à la maille régionale ou académique.

    Certaines routes, mentionnées par un cadenas, nécessitent une authentification. Pour obtenir une clef API, nous vous invitons à [contacter nos équipes](mailto:contact@inserjeunes.beta.gouv.fr).

    Le code source de l’API est public et disponible [sur Github](https://github.com/mission-apprentissage/trajectoires-pro/).

  license:
    name: MIT License
    url: https://raw.githubusercontent.com/mission-apprentissage/trajectoires-pro/main/LICENSE
  version: 1.0.0
paths:
...
  /api/inserjeunes/regionales.{ext}:
    get:
      description: Exporter des statistiques régionales
      security:
        - BearerAuth: [ ]
        - ApiKeyAuth: [ ]
      parameters:
        - $ref: "#/components/parameters/regions"
        - $ref: "#/components/parameters/millesimesRegional"
        - $ref: "#/components/parameters/code_certifications"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/items_par_page"
        - $ref: "#/components/parameters/ext"
      responses:
        200:
          $ref: "#/components/responses/RegionalesResponse"
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        404:
          $ref: "#/components/responses/404"
...
  /api/inserjeunes/regionales/{code_region}.{ext}:
    get:
      description: Exporter des statistiques d'une région
      security:
        - BearerAuth: [ ]
        - ApiKeyAuth: [ ]
      parameters:
        - $ref: "#/components/parameters/code_region"
        - $ref: "#/components/parameters/millesimesRegional"
        - $ref: "#/components/parameters/code_certifications"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/items_par_page"
        - $ref: "#/components/parameters/ext"
      responses:
        200:
          $ref: "#/components/responses/RegionalesResponse"
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        404:
          $ref: "#/components/responses/404"
...
components:
  parameters:
...
    millesimesRegional:
      name: millesimes
      in: query
      description: |
        Permet de filtrer la recherche avec un ou plusieurs millesimes séparés par des virgules.
        * Les millesimes au niveau régional sont des périodes (ex: 2020_2021)
        * Retourne les données du dernier millésime si le champ est vide
      schema:
        type: string
...
    code_certifications:
      name: code_certifications
      in: query
      description: |
        Permet de filtrer la recherche avec un ou plusieurs codes de certification séparés par des virgules ou des |.

        Un code certification peut-être :
          * un code formation diplome (CFD) pour la voie apprentissage
          * un code MEFSTAT11 pour la voie professionnelle
      schema:
        type: string
...
    regions:
      name: regions
      in: query
      schema:
        type: string
      description: |
        Permet de filtrer la recherche avec un ou plusieurs codes de région séparés par des virgules.
        
        Valeurs possibles :
         ```
         84 Auvergne-Rhône-Alpes
         27 Bourgogne-Franche-Comté
         53 Bretagne
         24 Centre-Val de Loire
         00 Collectivités d'outre-mer
         94 Corse
         44 Grand Est
         01 Guadeloupe
         03 Guyane
         32 Hauts-de-France
         04 La Réunion
         02 Martinique
         06 Mayotte
         28 Normandie
         75 Nouvelle-Aquitaine
         76 Occitanie
         52 Pays de la Loire
         93 Provence-Alpes-Côte d'Azur
         11 Île-de-France
         ```
...
    page:
      name: page
      in: query
      description: Le numéro de la page désirée
      schema:
        type: integer
    items_par_page:
      name: items_par_page
      in: query
      description: Le nombre maximum d'éléments dans la page
      schema:
        type: integer
        default: 10
...
    code_region:
      name: code_region
      in: path
      required: true
      schema:
        type: string
      description: |
        Un code region peut-être :
          * Le code INSEE d'une région
        
            Valeurs possibles :
            ```
            84 Auvergne-Rhône-Alpes
            27 Bourgogne-Franche-Comté
            53 Bretagne
            24 Centre-Val de Loire
            00 Collectivités d'outre-mer
            94 Corse
            44 Grand Est
            01 Guadeloupe
            03 Guyane
            32 Hauts-de-France
            04 La Réunion
            02 Martinique
            06 Mayotte
            28 Normandie
            75 Nouvelle-Aquitaine
            76 Occitanie
            52 Pays de la Loire
            93 Provence-Alpes-Côte d'Azur
            11 Île-de-France
            ```
          * Un code postal : dans ce cas la, l'API remontera elle même au code région.
...
    ext:
      name: ext
      in: path
      required: true
      schema:
        type: string
        enum: [ "json", "csv" ]
        default: json
  schemas:
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
        data:
          description: "Optionnel"
          type: object
          properties:
            millesimes:
              type: string
              description: "Millésime demandé (ou dernier millésime)"
            millesimesDisponible:
              type: array
              description: "Millésimes avec des données"
              items:
                type: string
    Pagination:
      type: object
      properties:
        page:
          type: string
        resultats_par_page:
          type: number
        nombre_de_page:
          type: number
        total:
          type: number
      description: Statut de la pagination
    Region:
      type: object
      required:
        - code
        - nom
      properties:
        code:
          type: string
          description: Le code INSEE de la région
        nom:
          type: string
          description: Le nom de la région
    Meta:
      type: object
      required:
        - date_import
      properties:
        date_import:
          type: string
        title:
          type: string
        description:
          type: string
        messages:
          description: Informations supplémentaires sur les données
          type: array
          items:
            type: string
          example: ["Les taux ne peuvent pas être affichés car il n'y a pas assez d'élèves pour fournir une information fiable."]
    DonneeSource:
      type: object
      required:
        - code
        - nom
      properties:
        code_certification:
          type: string
          description: Le code de la certification d'ou proviennent les données (Similaire dans le cas type = self)
        type:
          type: string
          description: |
            Le type de la source des données :
              * self : les données proviennent de la certification
              * ancienne : les données proviennent d'une certification plus ancienne
              * nouvelle : les données proviennent d'une certification plus récente
...
    Regionale:
      type: object
      description: |
        Statistiques d'une certification pour une région.

        Les taux retournés sont :
          * null quand les effectifs (nb_annee_term) sont < 20
          * absent si l'on ne peut pas les calculer (données manquantes)
      required:
        - millesime
        - code_certification
        - filiere
        - libelle
        - region
        - _meta
      properties:
        millesime:
          type: string
        code_certification:
          type: string
        filiere:
          type: string
          enum: [ "apprentissage", "pro" ]
        libelle:
          description: "Libellé de la certification"
          type: string
        libelle_ancien:
          description: "Ancien libellé de la certification (disponible en cas de changement de nom)"
          type: string
        donnee_source:
          $ref: "#/components/schemas/DonneeSource"
        nb_sortant:
          type: number
        nb_annee_term:
          type: number
        nb_poursuite_etudes:
          type: number
        nb_en_emploi_24_mois:
          type: number
        nb_en_emploi_18_mois:
          type: number
        nb_en_emploi_12_mois:
          type: number
        nb_en_emploi_6_mois:
          type: number
        taux_en_emploi_24_mois:
          type: number
          nullable: true
        taux_en_emploi_18_mois:
          type: number
          nullable: true
        taux_en_emploi_12_mois:
          type: number
          nullable: true
        taux_en_emploi_6_mois:
          type: number
          nullable: true
        taux_en_formation:
          type: number
          nullable: true
        taux_autres_6_mois:
          type: number
          nullable: true
        taux_autres_12_mois:
          type: number
          nullable: true
        taux_autres_18_mois:
          type: number
          nullable: true
        taux_autres_24_mois:
          type: number
          nullable: true
        region:
          $ref: "#/components/schemas/Region"
        _meta:
          $ref: "#/components/schemas/Meta"
...
  responses:
    '400':
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    '401':
      description: L'api key est manquante ou invalide
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RegionalesResponse:
      description: La liste des statistiques de certifications pour toutes les régions
      content:
        "application/json":
          schema:
            type: object
            properties:
              regionales:
                type: array
                items:
                  $ref: "#/components/schemas/Regionale"
              pagination:
                $ref: "#/components/schemas/Pagination"
        "text/csv; charset=UTF-8":
          schema:
            type: string
            example: |
              region;code_certification;filiere;millesime;donnee_source_type;donnee_source_code_certification;nb_annee_term;nb_en_emploi_12_mois;nb_en_emploi_18_mois;nb_en_emploi_24_mois;nb_en_emploi_6_mois;nb_poursuite_etudes;nb_sortant;taux_autres_12_mois;taux_autres_18_mois;taux_autres_24_mois;taux_autres_6_mois;taux_en_emploi_12_mois;taux_en_emploi_18_mois;taux_en_emploi_24_mois;taux_en_emploi_6_mois;taux_en_formation;taux_rupture_contrats
              Auvergne-Rhône-Alpes;01022103;apprentissage;2018_2019;self;01022103;106;55;48;55;46;38;67;12;19;12;21;52;45;52;43;36;
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: query
      name: apiKey
    BearerAuth:
      type: http
      scheme: bearer